Split function definitions
Split all funtion definitions from init files to do native compilations
for these defuns.

+++ setup.sh
@@ -15,14 +15,18 @@ sudo pacman -S --needed \
      clang pyright ruff
 
 # install sarasaTermSCNerd
-font_path=~/.local/share/fonts/
-if [[ ! -d ${font_path} ]]; then
-    mkdir -p ${font_path}
+fc-list | grep -i "Sarasa Term SC Nerd"
+if [[ $? -ne 0 ]]; then
+    font_path=~/.local/share/fonts/
+    if [[ ! -d ${font_path} ]]; then
+        mkdir -p ${font_path}
+    fi
+
+    wget https://github.com/laishulu/Sarasa-Term-SC-Nerd/releases/download/v2.3.1/SarasaTermSCNerd.ttc.tar.gz
+    tar xf ./SarasaTermSCNerd.ttc.tar.gz -C ${font_path}
+    sudo fc-cache
+    rm ./SarasaTermSCNerd.ttc.tar.gz
 fi
-wget https://github.com/laishulu/Sarasa-Term-SC-Nerd/releases/download/v2.3.1/SarasaTermSCNerd.ttc.tar.gz
-tar xf ./SarasaTermSCNerd.ttc.tar.gz -C ${font_path}
-sudo fc-cache
-rm ./SarasaTermSCNerd.ttc.tar.gz
 
 # setup python venv for lsp bridge
 lsp_bridge_pyenv_path=~/.emacs.d/lsp.bridge.pyenv
